Fix signed/unsigned comparisons in ViennaRNA/mfe/mfe.c

+++ src/ViennaRNA/mfe/mfe.c
@@ -588,7 +588,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
   Hgi = Hgj = Igi = Igj = Igp = Igq = Igg = Mgi = Mgj = 0;
 
   /* unfolded state */
-  eval = (hc->up_ext[1] >= (int)length) ? 1 : 0;
+  eval = (hc->up_ext[1] >= length) ? 1 : 0;
   if (hc->f)
     eval = (hc->f(1, length, 1, length, VRNA_DECOMP_EXT_UP, hc->data)) ? eval : 0;
 
@@ -643,7 +643,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
   for (j = MIN2(turn + 2, VRNA_GQUAD_MIN_BOX_SIZE); j <= length; j++) {
     /* regular base pairs */
     for (u = j - turn - 1; u >= 1; u--) {
-      eval = (hc->up_ml[1] >= (int)(u - 1)) ? (hc->mx[length * j + u] & VRNA_CONSTRAINT_CONTEXT_MB_LOOP) : 0;
+      eval = (hc->up_ml[1] >= (u - 1)) ? (hc->mx[length * j + u] & VRNA_CONSTRAINT_CONTEXT_MB_LOOP) : 0;
       if ((hc->f) && (!hc->f(1, j, u, j, VRNA_DECOMP_ML_ML, hc->data)))
         eval = 0;
 
@@ -676,7 +676,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
       /* g-quads */
       if (j >= VRNA_GQUAD_MIN_BOX_SIZE) {
         for (u = j - VRNA_GQUAD_MIN_BOX_SIZE + 1; u >= 1; u--) {
-          eval = (hc->up_ml[1] >= (int)(u - 1)) ? 1 : 0;
+          eval = (hc->up_ml[1] >= (u - 1)) ? 1 : 0;
           if ((hc->f) && (!hc->f(1, j, u, j, VRNA_DECOMP_ML_ML, hc->data)))
             eval = 0;
 
@@ -732,7 +732,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
             unsigned int u;
             u = i - j - 1;
             /* 1st, obey hard constraints */
-            if (hc->up_ext[j + 1] >= (int)u)
+            if (hc->up_ext[j + 1] >= u)
               eval = 1;
             if (hc->f)
               eval = (hc->f(j + 1, length, i, length, VRNA_DECOMP_EXT_EXT, hc->data)) ? eval : 0;
@@ -784,7 +784,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
           /* case 2.1: gquad forms an internal-loop like structure with another gquadruplex */
           for (u1 = 0, p = j + 1; p + VRNA_GQUAD_MIN_BOX_SIZE - 1 < i; p++, u1++) {
             /* obey hard constraints */
-            if (hc->up_int[j + 1] < (int)u1)
+            if (hc->up_int[j + 1] < u1)
               break;
 
             if (u1 > MAXLOOP)
@@ -792,7 +792,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
 
             for (u2 = 0, q = i - 1; q >= p + VRNA_GQUAD_MIN_BOX_SIZE - 1; q--, u2++) {
               /* obey hard constraints */
-              if (hc->up_int[q + 1] < (int)u2)
+              if (hc->up_int[q + 1] < u2)
                 break;
 
               if (u1 + u2 > MAXLOOP)
@@ -871,7 +871,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
           /* case 2.2: gquad forms an internal-loop like structure with another base pair */
           for (u1 = 0, p = j + 1; p + turn + 1 < i; p++, u1++) {
             /* obey hard constraints */
-            if (hc->up_int[j + 1] < (int)u1)
+            if (hc->up_int[j + 1] < u1)
               break;
 
             if (u1 > MAXLOOP)
@@ -879,7 +879,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
 
             for (u2 = 0, q = i - 1; q > p + turn; q--, u2++) {
               /* obey hard constraints */
-              if (hc->up_int[q + 1] < (int)u2)
+              if (hc->up_int[q + 1] < u2)
                 break;
               if (!(hc->mx[length * p + q] & (VRNA_CONSTRAINT_CONTEXT_INT_LOOP | VRNA_CONSTRAINT_CONTEXT_INT_LOOP_ENC)))
                 continue;
@@ -1001,9 +1001,9 @@ postprocess_circular(vrna_fold_compound_t *fc,
           if (u1 + u2 < 3)
             continue;
 
-          eval = (hc->up_ext[1] >= (int)u1) ? 1 : 0;
+          eval = (hc->up_ext[1] >= u1) ? 1 : 0;
           if (u2 > 0)
-            eval = (hc->up_ext[j + 1] >= (int)u2) ? eval : 0;
+            eval = (hc->up_ext[j + 1] >= u2) ? eval : 0;
           if (hc->f) {
             if (u1 > 0)
               eval = (hc->f(1, i - 1, 1, i - 1, VRNA_DECOMP_EXT_UP, hc->data)) ? eval : 0;
@@ -1074,7 +1074,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
     /* internal loop cases with at least one gquad below */
     for (i = 1; i < MAXLOOP + 1; i++) {
       u1 = i - 1;
-      if (hc->up_int[1] + 1 >= (int)i) {
+      if (hc->up_int[1] + 1 >= i) {
         /* [gquad] + (basepair) */
         for (j = i + VRNA_GQUAD_MIN_BOX_SIZE - 1; j + turn + 2 <= length; j++) {
 #ifndef VRNA_DISABLE_C11_FEATURES
@@ -1091,7 +1091,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
               if (stop + MAXLOOP < length + u1 + u2)
                 stop = length + u1 + u2 - MAXLOOP;
 
-              if (hc->up_int[j + 1] >= (int)u2) {
+              if (hc->up_int[j + 1] >= u2) {
                 for (u3 = 0, q = length; q >= stop; q--, u3++) {
                   if (((u2 == 0) && (u1 + u3 < 3)) ||
                       ((u1 + u3 == 0) && (u2 < 3)))
@@ -1099,7 +1099,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
                   unsigned int sq, sp;
                   sq = S1[q + 1];
                   sp = S1[p - 1];
-                  eval = (hc->up_int[q] >= (int)u3) ? 1 : 0;
+                  eval = (hc->up_int[q] >= u3) ? 1 : 0;
                   eval = (hc->mx[length * p + q] & (VRNA_CONSTRAINT_CONTEXT_INT_LOOP | VRNA_CONSTRAINT_CONTEXT_INT_LOOP_ENC)) ? eval : 0;
                   if (eval) {
                     int pq = indx[q] + p;
@@ -1160,7 +1160,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
           if (e != INF) {
             for (p = j + 1; p + VRNA_GQUAD_MIN_BOX_SIZE - 1 <= length; p++) {
               u2 = p - j - 1;
-              if (hc->up_int[j + 1] >= (int)u2) {
+              if (hc->up_int[j + 1] >= u2) {
                 unsigned int stop = p + VRNA_GQUAD_MIN_BOX_SIZE - 1;
 
                 if (stop + MAXLOOP < length + u1 + u2)
@@ -1242,7 +1242,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
 
             for (p = j + 1; p + VRNA_GQUAD_MIN_BOX_SIZE -1 <= length; p++) {
               u2 = p - j - 1;
-              if (hc->up_int[j + 1] < (int)u2)
+              if (hc->up_int[j + 1] < u2)
                 break;
 
               unsigned int stop = p + VRNA_GQUAD_MIN_BOX_SIZE - 1;
@@ -1255,7 +1255,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
                     ((u1 + u3 == 0) && (u2 < 3)))
                   continue;
 
-                eval = (hc->up_int[q] >= (int)u3) ? 1 : 0;
+                eval = (hc->up_int[q] >= u3) ? 1 : 0;
                 if (eval) {
 #ifndef VRNA_DISABLE_C11_FEATURES
                   int energy = vrna_smx_csr_get(c_gq, p, q, INF);
@@ -1976,7 +1976,7 @@ postprocess_circular(vrna_fold_compound_t *fc,
     } else if (FcM == Fc) {
       /* 1. find component in fM1_new */
       for (u = 1; u + MIN2(turn + 1, VRNA_GQUAD_MIN_BOX_SIZE - 1) <= Mi; u++) {
-        if (hc->up_ml[1] < (int)(u - 1))
+        if (hc->up_ml[1] < (u - 1))
           break;
 
         eval = ((hc->f) && (!hc->f(1, Mi, u, Mi, VRNA_DECOMP_ML_ML, hc->data))) ? 0 : 1;
